requirements.txt
streamlit
pandas
pillow
scikit-learn
numpy
openpyxl